temp_CAPI_Automation/requirements.txt
avocado-framework==82.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
jsonpath==0.82
lxml==4.9.1
paramiko==2.10.4
parsedatetime==2.6
pycparser==2.21
PyNaCl==1.5.0
pytils==0.3
pyvmomi==7.0.3
PyYAML==6.0
requests==2.27.1
scp==0.14.4
six==1.16.0
tools==0.1.9
urllib3==1.26.9
wait-for==1.2.0